cmd: Remove the `-i` option to `jj restore`.
It is superceded by the new options to the `jj diffedit` command.
